+++ scheduler-server/src/main/java/xyz/webspaghetti/schedulerserver/repository/UserRepository.java
@@ -5,7 +5,7 @@ import xyz.webspaghetti.schedulerserver.entity.User;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
-    Optional<User> findById(long id);
+    Optional<User> findById(Integer id);
 }
